Rename cursor_utils + add some comments
